pyproject.toml
[project]
name = "adk-scratchpad"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "googlemaps>=4.10.0",
    "google-adk>=1.26.0",
    "google-cloud-aiplatform>=1.140.0",
    "google-genai>=1.66.0",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.0",
    "scikit-learn>=1.8.0",
    "tqdm>=4.67.3",
    "uvicorn[standard]>=0.41.0",
    "websockets>=15.0.1",
    "rich>=13.0.0",
    "chainlit>=2.10.0",
]
